backend/requirements.txt
# FREESPORT Backend Dependencies
# Django 4.2 LTS + Django REST Framework 3.14+

# Core Django
Django==4.2.16
django-environ==0.11.2
requests==2.31.0
Pillow==10.2.0

# Django REST Framework
djangorestframework==3.14.0
django-cors-headers==4.3.1

# JWT Authentication
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Redis Cache
redis==5.0.1
django-redis==5.4.0

# API Documentation
drf-spectacular==0.28.0
django-filter==23.5

# Security & Validation
django-ratelimit==4.1.0
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Linting & Formatting
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Type Checking
mypy==1.7.1
django-stubs==4.2.6